requirements.txt
streamlit
pandas
plotly
pygwalker
sweetviz
numpy<2.0
google-cloud-bigquery
db-dtypes